fix: remove `resetTestSessions`
