+++ packages/serverpod_test/lib/serverpod_test_public_exports.dart
@@ -9,7 +9,6 @@ export 'serverpod_test.dart'
         ConnectionClosedException,
         flushMicrotasks,
         InvalidConfigurationException,
-        ResetTestSessions,
         RollbackDatabase,
         ServerpodInsufficientAccessException,
         ServerpodUnauthenticatedException,
