+++ tools/serverpod_cli/test/generator/dart/server_code_generator/test_tools_test.dart
@@ -64,7 +64,6 @@ void main() {
               r'withServerpod\(\n'
               r'  String testGroupName,\n'
               r'  _i\d\.TestClosure<TestEndpoints> testClosure, \{\n'
-              r'  _i\d\.ResetTestSessions\? resetTestSessions,\n'
               r'  _i\d\.RollbackDatabase\? rollbackDatabase,\n'
               r'  String\? runMode,\n'
               r'  bool\? enableSessionLogging,\n'
@@ -132,7 +131,6 @@ void main() {
               r'withServerpod\(\n'
               r'  String testGroupName,\n'
               r'  _i\d\.TestClosure<TestEndpoints> testClosure, \{\n'
-              r'  _i\d\.ResetTestSessions\? resetTestSessions,\n'
               r'  _i\d\.RollbackDatabase\? rollbackDatabase,\n'
               r'  String\? runMode,\n'
               r'  bool\? enableSessionLogging,\n'
@@ -178,7 +176,6 @@ void main() {
               r'withServerpod\(\n'
               r'  String testGroupName,\n'
               r'  _i\d\.TestClosure<TestEndpoints> testClosure, \{\n'
-              r'  _i\d\.ResetTestSessions\? resetTestSessions,\n'
               r'  _i\d\.RollbackDatabase\? rollbackDatabase,\n'
               r'  String\? runMode,\n'
               r'  bool\? enableSessionLogging,\n'
@@ -258,7 +255,6 @@ void main() {
               r'withServerpod\(\n'
               r'  String testGroupName,\n'
               r'  _i\d\.TestClosure<TestEndpoints> testClosure, \{\n'
-              r'  _i\d\.ResetTestSessions\? resetTestSessions,\n'
               r'  _i\d\.RollbackDatabase\? rollbackDatabase,\n'
               r'  String\? runMode,\n'
               r'  bool\? enableSessionLogging,\n'
